requirements.txt
cffi==1.17.1
colorthief==0.2.1
future==1.0.0
kociemba==1.2.1
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pycparser==2.22
RubikTwoPhase==1.1.1
scipy==1.14.1
setuptools==75.6.0
tk==0.1.0
